actually conformance not necessary
this doesn't allow multiple types to be nested (if they even can be)
however we don't need to check conformance because if the self type was
declared then it could have only been in a trait but the trait function
is available to us so it must've been because we implement that trait.

Need to do code gen and then some semantic checks to produce correct
error messages.